Add benchmark
From profiling data it seems like there isn't much we can do to optimize
the stack collapse sampler implementation. We could use Cython or wrap
some C/Rust, but it would open the pandora box of cross-compilation and
compatibility issues.

I think that for the moment this native Python implementation is good
enough, it takes on average ~3 μs to process a stack snapshot.

As for the impact on performances... `performance_impact.py` reports an
execution time with pylaprof lower than the native version.
The only explanation is that the impact of pylaprof is almost null and
that the lower execution times are caused by some arcane combination of
other factors.